fix(TAP-8111): Add validation to the time range of the statistical chart query. The default range has been set, and the logic for saving CPU memory indicators has been fixed

+++ manager/tm/src/main/java/com/tapdata/tm/worker/dto/MetricInfo.java
@@ -13,7 +13,7 @@ public class MetricInfo {
     /**
      * unit byte
      * */
-    Double heapMemoryUsage;
+    Long heapMemoryUsage;
 
     /**
      * unit %, such as 0.1 means 10%
